chore: migrate to @nrwl/workspace v11.1.2

+++ decorate-angular-cli.js
@@ -25,31 +25,12 @@ const fs = require('fs');
 const os = require('os');
 const cp = require('child_process');
 const isWindows = os.platform() === 'win32';
-const { output } = require('@nrwl/workspace');
-
-/**
- * Paths to files being patched
- */
-const angularCLIInitPath = 'node_modules/@angular/cli/lib/cli/index.js';
-
-/**
- * Patch index.js to warn you if you invoke the undecorated Angular CLI.
- */
-function patchAngularCLI(initPath) {
-  const angularCLIInit = fs.readFileSync(initPath, 'utf-8').toString();
-
-  if (!angularCLIInit.includes('NX_CLI_SET')) {
-    fs.writeFileSync(
-      initPath,
-      `
-if (!process.env['NX_CLI_SET']) {
-  const { output } = require('@nrwl/workspace');
-  output.warn({ title: 'The Angular CLI was invoked instead of the Nx CLI. Use "npx ng [command]" or "nx [command]" instead.' });
-}
-${angularCLIInit}
-    `
-    );
-  }
+let output;
+try {
+  output = require('@nrwl/workspace').output;
+} catch (e) {
+  console.warn('Angular CLI could not be decorated to enable computation caching. Please ensure @nrwl/workspace is installed.');
+  process.exit(0);
 }
 
 /**
@@ -65,32 +46,24 @@ function symlinkNgCLItoNxCLI() {
        * This is the most reliable way to create symlink-like behavior on Windows.
        * Such that it works in all shells and works with npx.
        */
-      ['', '.cmd', '.ps1'].forEach((ext) => {
-        if (fs.existsSync(nxPath + ext))
-          fs.writeFileSync(ngPath + ext, fs.readFileSync(nxPath + ext));
+      ['', '.cmd', '.ps1'].forEach(ext => {
+        if (fs.existsSync(nxPath + ext)) fs.writeFileSync(ngPath + ext, fs.readFileSync(nxPath + ext));
       });
     } else {
       // If unix-based, symlink
       cp.execSync(`ln -sf ./nx ${ngPath}`);
     }
-  } catch (e) {
-    output.error({
-      title:
-        'Unable to create a symlink from the Angular CLI to the Nx CLI:' +
-        e.message,
-    });
+  }
+  catch(e) {
+    output.error({ title: 'Unable to create a symlink from the Angular CLI to the Nx CLI:' + e.message });
     throw e;
   }
 }
 
 try {
   symlinkNgCLItoNxCLI();
-  patchAngularCLI(angularCLIInitPath);
-  output.log({
-    title: 'Angular CLI has been decorated to enable computation caching.',
-  });
-} catch (e) {
-  output.error({
-    title: 'Decoration of the Angular CLI did not complete successfully',
-  });
+  require('@nrwl/cli/lib/decorate-cli').decorateCli();
+  output.log({ title: 'Angular CLI has been decorated to enable computation caching.' });
+} catch(e) {
+  output.error({ title: 'Decoration of the Angular CLI did not complete successfully' });
 }
